% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap.R
\name{is_leap}
\alias{is_leap}
\title{Check if number given is a leap year}
\usage{
is_leap(year)
}
\arguments{
\item{year}{Year to input}
}
\value{
Boolean
}
\description{
Check if number given is a leap year
}
\examples{
is_leap(2020)
#> TRUE

}
